+++ packages/visibility_detector/lib/src/visibility_detector_layer.dart
@@ -18,7 +18,7 @@ import 'visibility_detector_controller.dart';
 /// ancestors.  The returned sequence is in [parent, child] order.
 Iterable<Layer> _getLayerChain(Layer start) {
   final layerChain = <Layer>[];
-  for (var layer = start; layer != null; layer = layer.parent) {
+  for (Layer? layer = start; layer != null; layer = layer.parent) {
     layerChain.add(layer);
   }
   return layerChain.reversed;
@@ -61,10 +61,10 @@ Rect _localRectToGlobal(Layer layer, Rect localRect) {
 class VisibilityDetectorLayer extends ContainerLayer {
   /// Constructor.  See the corresponding properties for parameter details.
   VisibilityDetectorLayer(
-      {@required this.key,
-      @required this.widgetSize,
-      @required this.paintOffset,
-      @required this.onVisibilityChanged})
+      {required this.key,
+      required this.widgetSize,
+      required this.paintOffset,
+      required this.onVisibilityChanged})
       : assert(key != null),
         assert(paintOffset != null),
         assert(widgetSize != null),
@@ -72,16 +72,21 @@ class VisibilityDetectorLayer extends ContainerLayer {
         _layerOffset = Offset.zero;
 
   /// Timer used by [_scheduleUpdate].
-  static Timer _timer;
+  static Timer? _timer;
 
-  /// Keeps track of [VisibilityDetectorLayer] objects that have been recently
-  /// updated and that might need to report visibility changes.
-  ///
-  /// Additionally maps [VisibilityDetector] keys to the most recently added
+  /// Maps [VisibilityDetector] keys to the most recently added
   /// [VisibilityDetectorLayer] that corresponds to it; this mapping is
   /// necessary in case a layout change causes a new layer to be instantiated
   /// for an existing key.
-  static final _updated = <Key, VisibilityDetectorLayer>{};
+  static final _activeLayers = <Key, VisibilityDetectorLayer>{};
+
+  /// Tracks keys for [VisibilityDetectorLayer]s that have been recently updated
+  /// and that might need to report visibility changes.
+  static final _updatedKeys = <Key>{};
+
+  /// Tracks keys for [VisibilityDetectorLayer]s that have pending callbacks
+  /// requested via [VisibilityDetectorController.scheduleNotification].
+  static final _forcedKeys = <Key>{};
 
   /// Keeps track of the last known visibility state of a [VisibilityDetector].
   ///
@@ -96,27 +101,20 @@ class VisibilityDetectorLayer extends ContainerLayer {
   static final _lastBounds = <Key, Rect>{};
 
   /// The key for the corresponding [VisibilityDetector] widget.
-  ///
-  /// Never null.
   final Key key;
 
   /// The size of the corresponding [VisibilityDetector] widget.
-  ///
-  /// Never null.
   final Size widgetSize;
 
-  /// Last known layer offset supplied to [addToScene].  Never null.
+  /// Last known layer offset supplied to [addToScene].
   Offset _layerOffset;
 
-  /// The offset supplied to [RenderVisibilityDetector.paint] method.  Never
-  /// null.
+  /// The offset supplied to [RenderVisibilityDetector.paint] method.
   final Offset paintOffset;
 
   /// See [VisibilityDetector.onVisibilityChanged].
   ///
   /// Do not invoke this directly; call [_fireCallback] instead.
-  ///
-  /// Never null.
   final VisibilityChangedCallback onVisibilityChanged;
 
   /// Computes the bounds for the corresponding [VisibilityDetector] widget, in
@@ -129,17 +127,17 @@ class VisibilityDetectorLayer extends ContainerLayer {
   /// Computes the accumulated clipping bounds, in global coordinates.
   Rect _computeClipRect() {
     assert(RendererBinding.instance?.renderView != null);
-    var clipRect = Offset.zero & RendererBinding.instance.renderView.size;
+    var clipRect = Offset.zero & RendererBinding.instance!.renderView.size;
 
     var parentLayer = parent;
     while (parentLayer != null) {
-      Rect curClipRect;
+      Rect? curClipRect;
       if (parentLayer is ClipRectLayer) {
         curClipRect = parentLayer.clipRect;
       } else if (parentLayer is ClipRRectLayer) {
-        curClipRect = parentLayer.clipRRect.outerRect;
+        curClipRect = parentLayer.clipRRect!.outerRect;
       } else if (parentLayer is ClipPathLayer) {
-        curClipRect = parentLayer.clipPath.getBounds();
+        curClipRect = parentLayer.clipPath!.getBounds();
       }
 
       if (curClipRect != null) {
@@ -157,22 +155,27 @@ class VisibilityDetectorLayer extends ContainerLayer {
     return clipRect;
   }
 
+  /// Marks the [VisibilityDetectorLayer] as being recently updated and
+  /// schedules a timer to invoke visibility callbacks.
+  void _scheduleUpdate() {
+    _updatedKeys.add(key);
+    _scheduleCallbacks();
+  }
+
   /// Schedules a timer to invoke the visibility callbacks.  The timer is used
   /// to throttle and coalesce updates.
-  void _scheduleUpdate() {
-    final isFirstUpdate = _updated.isEmpty;
-    _updated[key] = this;
-
+  static void _scheduleCallbacks() {
     final updateInterval = VisibilityDetectorController.instance.updateInterval;
     if (updateInterval == Duration.zero) {
       // Even with [Duration.zero], we still want to defer callbacks to the end
       // of the frame so that they're processed from a consistent state.  This
       // also ensures that they don't mutate the widget tree while we're in the
       // middle of a frame.
+      final isFirstUpdate = _updatedKeys.length + _forcedKeys.length == 1;
       if (isFirstUpdate) {
         // We're about to render a frame, so a post-frame callback is guaranteed
         // to fire and will give us the better immediacy than `scheduleTask<T>`.
-        SchedulerBinding.instance.addPostFrameCallback((timeStamp) {
+        SchedulerBinding.instance!.addPostFrameCallback((timeStamp) {
           _processCallbacks();
         });
       }
@@ -181,7 +184,7 @@ class VisibilityDetectorLayer extends ContainerLayer {
       // to the update duration will be picked up automatically.
       _timer = Timer(updateInterval, _handleTimer);
     } else {
-      assert(_timer.isActive);
+      assert(_timer!.isActive);
     }
   }
 
@@ -195,7 +198,7 @@ class VisibilityDetectorLayer extends ContainerLayer {
     // of `addPostFrameCallback` or `scheduleFrameCallback` so that work will
     // be done even if a new frame isn't scheduled and without unnecessarily
     // scheduling a new frame.
-    SchedulerBinding.instance
+    SchedulerBinding.instance!
         .scheduleTask<void>(_processCallbacks, Priority.touch);
   }
 
@@ -206,14 +209,20 @@ class VisibilityDetectorLayer extends ContainerLayer {
     _processCallbacks();
   }
 
+  /// See [VisibilityDetectorController.scheduleNotification].
+  static void scheduleNotification(Key key) {
+    _forcedKeys.add(key);
+    _scheduleCallbacks();
+  }
+
   /// Removes entries corresponding to the specified [Key] from our internal
   /// caches.
   static void forget(Key key) {
-    _updated.remove(key);
+    _updatedKeys.remove(key);
     _lastVisibility.remove(key);
     _lastBounds.remove(key);
 
-    if (_updated.isEmpty) {
+    if (_updatedKeys.isEmpty && _forcedKeys.isEmpty) {
       _timer?.cancel();
       _timer = null;
     }
@@ -222,34 +231,51 @@ class VisibilityDetectorLayer extends ContainerLayer {
   /// Executes visibility callbacks for all updated [VisibilityDetectorLayer]
   /// instances.
   static void _processCallbacks() {
-    for (final layer in _updated.values) {
-      if (!layer.attached) {
-        layer._fireCallback(VisibilityInfo(
-            key: layer.key, size: _lastVisibility[layer.key]?.size));
+    for (final key in _forcedKeys) {
+      final layer = _activeLayers[key];
+      if (layer == null) {
         continue;
       }
-
-      final widgetBounds = layer._computeWidgetBounds();
-      _lastBounds[layer.key] = widgetBounds;
-
-      final info = VisibilityInfo.fromRects(
-          key: layer.key,
-          widgetBounds: widgetBounds,
-          clipRect: layer._computeClipRect());
-      layer._fireCallback(info);
+      _updatedKeys.remove(layer.key);
+      layer._fireCallback(force: true);
     }
-    _updated.clear();
+    _forcedKeys.clear();
+
+    for (final key in _updatedKeys) {
+      final layer = _activeLayers[key]!;
+      layer._fireCallback(force: false);
+    }
+    _updatedKeys.clear();
   }
 
   /// Invokes the visibility callback if [VisibilityInfo] hasn't meaningfully
-  /// changed since the last time we invoked it.
-  void _fireCallback(VisibilityInfo info) {
-    assert(info != null);
+  /// changed since the last time we invoked it or if callback invocation is
+  /// forced.
+  void _fireCallback({required bool force}) {
+    late VisibilityInfo info;
+
+    if (!attached) {
+      _activeLayers.remove(key);
+      info = VisibilityInfo(
+        key: key,
+        size: _lastVisibility[key]?.size,
+      );
+    } else {
+      final widgetBounds = _computeWidgetBounds();
+      _lastBounds[key] = widgetBounds;
+      info = VisibilityInfo.fromRects(
+        key: key,
+        widgetBounds: widgetBounds,
+        clipRect: _computeClipRect(),
+      );
+    }
 
     final oldInfo = _lastVisibility[key];
     final visible = !info.visibleBounds.isEmpty;
 
-    if (oldInfo == null) {
+    if (force) {
+      // Don't filter it out.
+    } else if (oldInfo == null) {
       if (!visible) {
         return;
       }
@@ -279,6 +305,7 @@ class VisibilityDetectorLayer extends ContainerLayer {
   /// See [AbstractNode.attach].
   @override
   void attach(Object owner) {
+    _activeLayers[key] = this;
     super.attach(owner);
     _scheduleUpdate();
   }